refactor extractStyleProps
