+++ src/babel-plugin/extractStyleProps/fromObject.ts
@@ -1,87 +1,7 @@
 import { types as t, type NodePath } from "@babel/core";
-import type { JSXOpeningElement } from "@babel/types";
-import { isStyledProp } from "../system";
+import { isStyledProp } from "../../system";
 import { PseudoProps, isPseudoProps } from "src/system/pseudo";
-import { CSSProperty } from "../system/types";
-
-type ExtractedStyleProps<T> = T extends NodePath<t.JSXOpeningElement>
-  ? {
-      filteredAttributes: t.JSXAttribute[];
-      styledProps: {
-        [key: string]: CSSProperty | number | (CSSProperty | number)[];
-      };
-      pseudoProps: PseudoProps;
-    }
-  : {
-      filteredProperties: t.ObjectProperty[];
-      styledProps: { [key: string]: string | number | (string | number)[] };
-      pseudoProps: PseudoProps;
-    };
-
-/**
- * Extracts style props from a JSX opening element or an ObjectExpression in a React.createElement call
- * and returns the filtered properties and the extracted style props.
- */
-export function extractStyleProps<
-  T extends NodePath<t.JSXOpeningElement> | NodePath<t.ObjectExpression>
->(path: T): ExtractedStyleProps<T> {
-  if (t.isJSXOpeningElement(path.node)) {
-    return extractStylePropsFromAST(path.node) as ExtractedStyleProps<T>;
-  } else {
-    return extractStylePropsFromObjectExpression(
-      path as NodePath<t.ObjectExpression>,
-      path.node
-    ) as ExtractedStyleProps<T>;
-  }
-}
-
-export function extractStylePropsFromAST(
-  openingElement: JSXOpeningElement
-): ExtractedStyleProps<NodePath<t.JSXOpeningElement>> {
-  const styledProps: { [key: string]: string | number | (string | number)[] } =
-    {};
-  const pseudoProps: PseudoProps = {};
-
-  const filteredAttributes = openingElement.attributes.filter((attr) => {
-    if (
-      t.isJSXAttribute(attr) &&
-      t.isJSXIdentifier(attr.name) &&
-      isStyledProp(attr.name.name)
-    ) {
-      if (t.isStringLiteral(attr.value)) {
-        styledProps[attr.name.name] = attr.value.value;
-      } else if (t.isJSXExpressionContainer(attr.value)) {
-        const { expression } = attr.value;
-        if (t.isStringLiteral(expression) || t.isNumericLiteral(expression)) {
-          styledProps[attr.name.name] = expression.value;
-        } else if (t.isArrayExpression(expression)) {
-          styledProps[attr.name.name] = expression.elements
-            .map((e) => {
-              if (e?.type === "NumericLiteral" || e?.type === "StringLiteral") {
-                return e.value;
-              }
-            })
-            .filter(Boolean) as (string | number)[];
-        }
-      }
-      return false;
-    } else if (
-      t.isJSXAttribute(attr) &&
-      t.isJSXIdentifier(attr.name) &&
-      t.isObjectExpression(attr.value) &&
-      isPseudoProps(attr.name.name)
-    ) {
-      Object.assign(pseudoProps, {
-        ...pseudoProps,
-        [attr.name.name]: extractStylePropsFromAST(attr.value).styledProps,
-      });
-      return false;
-    }
-    return true;
-  }) as t.JSXAttribute[];
-
-  return { filteredAttributes, styledProps, pseudoProps };
-}
+import { ExtractedStyleProps } from ".";
 
 export function extractStylePropsFromObjectExpression(
   path: NodePath<t.ObjectExpression>,
