+++ src/babel-plugin/processHTMLTag.ts
@@ -1,9 +1,6 @@
 import { NodePath, PluginObj, types as t } from "@babel/core";
 import { CallExpression, JSXElement, JSXOpeningElement } from "@babel/types";
-import {
-  extractStylePropsFromAST,
-  extractStylePropsFromObjectExpression,
-} from "./extractStyleFromAST";
+import { extractStyleProps } from "./extractStyleFromAST";
 import { sheet } from "../sheet";
 import { all } from "../system";
 import { PseudoProps, pseudoMappings } from "src/system/pseudo";
@@ -27,7 +24,7 @@ const processJSXHTMLTag = (path: NodePath<t.JSXOpeningElement>) => {
 
   if (dataAttribute) return;
   const { filteredAttributes, styledProps, pseudoProps } =
-    extractStylePropsFromAST(path.node);
+    extractStyleProps(path);
   // Update the attributes of the opening element by removing the styled props,
   // so that the styled props don't get passed down as regular HTML attributes.
   path.node.attributes = filteredAttributes;
@@ -58,7 +55,7 @@ const processReactCreateElementHTMLTag = (
   path: NodePath<t.ObjectExpression>
 ) => {
   const { filteredProperties, styledProps, pseudoProps } =
-    extractStylePropsFromObjectExpression(path, path.node);
+    extractStyleProps(path);
   // Update the properties of the object expression by removing the styled props,
   // so that the styled props don't get passed down as regular HTML attributes.
   path.node.properties = filteredProperties;
